requirements.txt
mistralai==1.7.0
python-dotenv==1.1.0